# apps/backend/swagger.json
{"openapi":"3.0.0","paths":{"/ping":{"get":{"operationId":"AppController_get","summary":"Ping","description":"動作確認用です。","parameters":[],"responses":{"200":{"description":""}},"tags":["Ping"]}},"/health":{"get":{"operationId":"AppController_health","summary":"Health","description":"ヘルスチェック用です。","parameters":[],"responses":{"200":{"description":""}},"tags":["Ping"]}},"/users":{"get":{"operationId":"UserController_list","summary":"ユーザー一覧を取得","parameters":[{"name":"page","required":false,"in":"query","description":"ページネーション: ページ番号","schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","description":"ページネーション: 1ページのデータ数","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicListUsersResponse"}}}}},"tags":["PublicService"]},"post":{"operationId":"UserController_create","summary":"ユーザーを新規作成","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicCreateUserRequestBody"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicCreateUserResponse"}}}}},"tags":["PublicService"]}},"/users/{id}":{"delete":{"operationId":"UserController_delete","summary":"ユーザーの削除","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["PublicService"]}},"/signatures":{"get":{"operationId":"SignatureController_list","summary":"署名一覧を取得","parameters":[{"name":"page","required":false,"in":"query","description":"ページネーション: ページ番号","schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","description":"ページネーション: 1ページのデータ数","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicListSignaturesResponse"}}}}},"tags":["PublicService"]},"post":{"operationId":"SignatureController_create","summary":"署名を新規作成","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicCreateSignatureRequestBody"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicCreateSignatureResponse"}}}}},"tags":["PublicService"]}},"/signatures/{id}":{"delete":{"operationId":"SignatureController_delete","summary":"署名の削除","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["PublicService"]}},"/web3/balance":{"get":{"operationId":"Web3Controller_getBalance","parameters":[{"name":"chainId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["PublicService"]}},"/web3/unconfirmed_transactions":{"get":{"operationId":"Web3Controller_listTransactions","parameters":[{"name":"contractAddress","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["PublicService"]}},"/greeter":{"get":{"operationId":"GreeterController_get","parameters":[],"responses":{"200":{"description":""}},"tags":["PublicService"]},"post":{"operationId":"GreeterController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["PublicService"]}},"/greeter/{hash}":{"get":{"operationId":"GreeterController_getTransactionConfirmations","parameters":[{"name":"hash","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["PublicService"]}},"/factory/deploy":{"post":{"operationId":"FactoryController_deploy","parameters":[],"responses":{"201":{"description":""}},"tags":["PublicService"]}},"/market/transfer":{"post":{"operationId":"MarketController_transfer","parameters":[],"responses":{"201":{"description":""}},"tags":["PublicService"]}},"/erc721/count":{"get":{"operationId":"ERC721Controller_count","parameters":[{"name":"address","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["PublicService"]}},"/erc721/mint":{"post":{"operationId":"ERC721Controller_create","parameters":[],"responses":{"201":{"description":""}},"tags":["PublicService"]}}},"info":{"title":"NFT Market","description":"NFT Market API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"User":{"type":"object","properties":{"createdAt":{"type":"number","description":"データ作成日時を表すUnixTime（秒）","example":1610377347},"updatedAt":{"type":"number","description":"データ更新日時を表すUnixTime（秒）","example":1610377347},"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"}},"required":["createdAt","updatedAt","id","name","email"]},"Pagination":{"type":"object","properties":{"totalCount":{"type":"number","description":"ヒット件数","example":45},"offsetValue":{"type":"number","description":"オフセット","example":15},"perPage":{"type":"number","description":"1ページのデータ数","example":10},"totalPages":{"type":"number","description":"ページ総数","example":7},"currentPage":{"type":"number","description":"現在のページ","example":3},"nextPage":{"type":"number","description":"次のページ","example":4},"prevPage":{"type":"number","description":"前のページ","example":2},"isFirstPage":{"type":"boolean","description":"最初のページ","example":1},"isLastPage":{"type":"boolean","description":"最後のページ","example":1}},"required":["totalCount","perPage","currentPage"]},"PublicListUsersResponse":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["users","pagination"]},"PublicCreateUserRequestBody":{"type":"object","properties":{"address":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["address","name","email","password"]},"PublicCreateUserResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"}},"required":["user"]},"Signature":{"type":"object","properties":{"createdAt":{"type":"number","description":"データ作成日時を表すUnixTime（秒）","example":1610377347},"updatedAt":{"type":"number","description":"データ更新日時を表すUnixTime（秒）","example":1610377347},"id":{"type":"string"},"sign":{"type":"string"},"deadline":{"type":"number"},"tokenId":{"type":"number"},"chainId":{"type":"number"}},"required":["createdAt","updatedAt","id","sign","deadline","tokenId","chainId"]},"PublicListSignaturesResponse":{"type":"object","properties":{"signatures":{"type":"array","items":{"$ref":"#/components/schemas/Signature"}},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["signatures","pagination"]},"PublicCreateSignatureRequestBody":{"type":"object","properties":{"sign":{"type":"string"},"deadline":{"type":"number"},"tokenId":{"type":"number"},"chainId":{"type":"number"},"nonce":{"type":"number"},"address":{"type":"string"}},"required":["sign","deadline","tokenId","chainId","nonce","address"]},"PublicCreateSignatureResponse":{"type":"object","properties":{"signature":{"$ref":"#/components/schemas/Signature"}},"required":["signature"]}}}}